Upgrades GODMODE from a flag to a trait. (#5276)
## About The Pull Request

Ports https://github.com/tgstation/tgstation/pull/86596 and
https://github.com/tgstation/tgstation/pull/86718

> GODMODE has a lot of sources that toggle it. From admin-stuff to
status effects, components, actions and mobs which are supposed to be
invincible. It's better off as a trait than a flag, so we can manage
these sources.

## Why It's Good For The Game

> See above.

## Changelog
:cl: Absolucy, Ghommie, SmArtKar
admin: godmode is now a datum trait instead of a bitflag. This means the
process for toggling it is a little different now.
/:cl: