Syndicate Depot Updates and Shipbreaking Is Now Ruin-Friendly (#5246)
## About The Pull Request

- The Syndicate Depot has had its interior security cameras removed and
its airlocks and APCs modified to keep nosy AIs from fucking with
equipment. Johnson and Co refuses to admit any wrongdoing.
- The Depot has also been fitted with shipbreaking, both as a proof of
concept to add to other ruins and because it's something for the techies
to do.
- All shipbreaking templates have been modified to only have passthrough
areas, as otherwise it would break depot shipbreaking (and ruin
shipbreaking in general in the future)
- The depot's circuit imprinter has also been replaced with an ancient
one to mirror some other PR that I don't have the number on hand for
right now so that they can sod off from the depot to prevent merge
conflicts

## Why It's Good For The Game

Shipbreaking is fucking cool, and extending access to it to ruins is
good.

Also, AIs fucking with the depot is not good.

## Changelog
:cl:
add: The Syndicate depot now has shipbreaking!
add: Shipbreaking can now also be integrated into other space ghostroles
with enough effort! (good luck lmao)
balance: The Syndicate depot's circuit imprinter is now the shit ancient
one, but they have been provided quantum pad boards so they can still be
buddy-buddy with other outposts.
fix: Nanotrasen AIs can no longer fuck with the syndicate depot, at
least on the interior. They can still see the outside, but it's now much
harder for them to do anything.
/:cl: