Custom Shoe Sounds fix. (#5388)
## About The Pull Request

Elements and components aren't easily explained. I made an oops thinking
I could make this an element. Caused anything using this element to
share the same settings. This changes it to a component and fixes this
"quantum locking" issue.

Fixes #5365

## Why It's Good For The Game

Shoe sounds are supposed to be instanced on each shoe its applied to not
shared between each shoe. Defeats the purpose of custom shoe sounds.
This refactor fixes it.

## Changelog
:cl: Siro
fix: Fixed all jackboots being quantum entangled - toggling the
noisemaker for one pair of jackboots no longer changes it for every
other pair of jackboots too.
/:cl: