[PORT] Two new brain traumas; plus, a rewritten Kleptomania (#5138)
## About The Pull Request
This PR ports the following PR:
* https://github.com/tgstation/tgstation/pull/82129

and additionally ports a fix for the above PR:
* https://github.com/tgstation/tgstation/pull/82741

This is a summary of the list of changes:
* Adds a new Mild-type brain trauma, "Possessive."
* Occasionally, you'll become possessive of an item in a random hand
slot foor a short time. This prevents you from dropping the item or
moving it to another slot.
* Adds a new Special-type brain trauma, "Primal Instincts."
* Your character will occasionally be taken over by a monkey AI. You
will still be in control, but you'll have to fight the monkey AI.
  * There is a high chance the monkey AI will be aggressive.
* The monkey AI *does not* care what you're trying to do. It will drop
items, take wearables off, and so much more. Woe be upon you if you get
this brain trauma.
* On the bright side, you can understand monkey language while your
primal instincts are active.
* Replaces our current Kleptomania with a new Kleptomania.
* The chance for Kleptomania to activate is higher than before, but also
has a 8-second cooldown every time it activates. It also won't activate
if you've taken damage recently.
* Before, you could only ever steal to your active hand - and as a
result, your active hand had to be free for it to activate. Now, it can
activate if *any* hand is free.
* Kleptomaniacs now steal a random item, instead of the first one they
see.
* Kleptomaniacs who are pickpocketing now steal from a random valid
victim, instead of the first one they see.

## Why It's Good For The Game
Cool and interesting drain bramage mechanics!

## New admin logs
The new traumas will generate a number of logs under the following
circumstances:
* When Possessive causes the owner to become (un)possessive of an item;
located under the ATTACK logs, in orange
* When Primal Instincts is activated/deactivated; located under the
ATTACK logs, in orange
* When a Kleptomaniac attempts to pick up an item off the ground;
located under the ATTACK logs, in orange
* When a Kleptomaniac attempts (or succeeds) a pickpocket; located under
the ATTACK logs, in red
* When someone is the victim of a Kleptomaniac pickpocket; located under
the VICTIM logs, in orange

## Changelog

:cl:MichiRecRoom, MrMelbert
add: Adds a new Mild-type brain trauma, Possessive. Occasionally, an
item you're holding in your hand will become undroppable for a short
time.
add: Adds a new Special-type brain trauma, Primal Instincts.
Occasionally, your character will be taken over by a (possibly
aggressive) monkey AI.
add: Rewrites our Kleptomania atop the new Kleptomania. This brings a
number of changes, including activating more frequently (but with a
cooldown), and the ability for the trauma to steal to any available
hand.
admin: New logs for the new brain traumas. See PR 5138 for the details
of these logs, and where to find them.
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: Lucy <lucy@absolucy.moe>