Remove a period from the regex used to determine valid youtube IDs (#5422)
## About The Pull Request
Does what it says on the tin.

In the context of its location in this regex, this period matches a
literal period - `.` - which means periods are allowed to be in youtube
IDs. However periods can never appear in youtube IDs - or at least, I've
never seen them. So there's no reason for this regex to allow them
through.

## Why It's Good For The Game
Tiny optimization, I guess?

## Changelog
No player-facing changes - not unless they were submitting invalid
youtube IDs?