ERT Code Yellow (#5563)
## About The Pull Request
Adds a yellow alert ERT into the game.

Also access locks Real Death Ripley's to Specops access
## Why It's Good For The Game
People have believed a engineering ERT exists when it does not, now it
does.

Akin to Code Purple the leader is a blue ERT while the rest are generic
ERT, all of the engineering variety.

They come equipped with their own dropship, with a custom made Ripley
with a clamp and RCD. Being the almost the best of both worlds of the
Ripley MK1 and MK2. A spacial nullifier, some floorbots, and enough
materials to secure a decent portion of the station on their own
## Changelog
:cl:
add: ERT Code: Yellow has been initialized
fix: Real Death Ripleys will have SPECOPS access. As normal civilians
should not pilot nor know what the existence of anything I said is.
/:cl: