small changes to abandoned syndicate infiltrator (#5505)
## About The Pull Request

small changes to the abandoned syndicate infiltrator
you can now place an apc on it
it now has it's own area
slightly extends the abandoned infiltrator area to cover the missing
sections of the ship because making areas is annoying

## Why It's Good For The Game

not being able to place an apc on the abandoned infiltrator is stupid

## Changelog

:cl:
add: The abandoned infiltrator now has it's own area.
fix: You can now place an APC on the abandoned infiltrator.
fix: The blast door button on the abandoned infiltrator now requires
syndicate access.
qol: The airlock electronics on the abandoned infiltrator are now preset
with the name "Medical" and syndicate access.
qol: Moves the walls between the cockpit and hallway on the abandoned
infiltrator up by one tile for consistency with the version of the
infiltrator it was based on.
qol: Extends the abandoned infiltrator area to cover the missing
sections of the ship.
/:cl: