Refactors silicon emotes slightly - port of tgstation 84912 (#5384)
## About The Pull Request

There was a small bug - creatures with synthetic voice boxes (IPCs,
people with robot voicebox) were unable to use *beep due to a strange
configuration of can_run_emote with monke modularity. I took a look at
it, and it seemed like a good time to port this PR:
https://github.com/tgstation/tgstation/pull/84912

This essentially moves the check to a trait, and that trait is added to
silicons, synthetic tongue organs, bots, and brains (posibrain/MMIs).

## Why It's Good For The Game

Code's a bit cleaner in terms of defining who can use silicon emotes,
and this fixes a bug with using *beep as a carbon with a synthetic voice
box.

## Changelog
:cl:
refactor: Refactored the way silicon emotes are handled in silicons,
robots, and synthetic voiceboxes
fix: synthetic voiceboxes can *beep again!
/:cl:

Co-authored-by: Ziiro <thechaosofzero@gmail.com>