Lone Op weight changes (#5512)
## About The Pull Request
Changes it so lone operatives have a minimum weight of 1, so that there
is always a slim chance for them to spawn if the antag cap hasn't been
reached (for reference, this is the same chance as a slaughter demon
spawning). Changes it so that weight will start to attempt to increase
after 5 minutes of the disk being unsecure, down from 8.3.
## Why It's Good For The Game
Right now, lone ops rarely, if ever, spawn in. This is partly due to the
fact that it takes about 20 minutes of the disk remaining unsecure for
lone ops to bypass the antag cap, and partly due to it taking almost 12
minutes in order for lone ops to even have a chance to spawn at all.
This will hopefully make it so leaving the disk unsecure will actually
result in a lone op spawning, and give lone ops a slim chance to spawn
even if the disk is secure.
## Changelog
:cl:
balance: Reduced the time it takes for an unsecure disk to start
increasing lone op weight (5 minutes, down from 8.3)
balance: Lone ops now have a minimum weight of 1
/:cl: