Fixes RDS quirk not working on non-IPCs (#5589)
## About The Pull Request

RDS was broken by #4992 due to trying to replacing the add() proc of
insanity quirk instead of adding lines to it... Ideally IPCs wouldn't
have to go through the entire RDS quirk adding process but for now this
will do.
## Why It's Good For The Game

Fixes #5311
Reality Dissociation Syndrome was broken for a while and was a free 8
quirk points, this fixes it.

## Changelog
:cl:
fix: Reality Dissociation Syndrome properly works now.
/:cl: