Caffeine dependence quirk (#5623)
## About The Pull Request
Adds a single -2 point quirk to the game.
You will now be addicted to coffee and have to keep drinking it, similar
to the other addiction quirks.
Adds the needed coffee addiction to facilitate the quirk, a somewhat
milder version of the stimulants addiction already in the game.


## Why It's Good For The Game

For all those caffeine addicts out there.
And a bit more character customisation

## Changelog

:cl:
add: Adds the Caffeine dependence quirk. Get to drinking that coffee.
add: Makes coffee mildly addictive.
add: Coffee addiction.

/:cl: