Reduces the range of flashbang guaranteed stun to <1 (#5607)
## About The Pull Request

Simply makes flashbangs only get a stun on protected individuals if they
are on the SAME tile as the flashbang. No more good cop / suicide cop
dynamic duos (one guy runs in and stands next to bad guy with grenade in
hand and the other follows with batong).


## Why It's Good For The Game

Who thought giving security a guaranteed AOE hard stun was a good idea?

## Changelog


:cl:

balance: Flashbangs only get a guaranteed stun on the same tile, not
adjacent tiles.
/:cl: