Temperature Guns Fun (#5582)
## About The Pull Request


## Why It's Good For The Game
We have both the thermal guns and the temperature guns already in the
game.
In their current state neither of them are ever used as the math behind
the scenes are way outdated.
This will bring them up to speed and make them actually useful now.

## Changelog
Alters how the temperature calculations are done for the temperature
guns to ensure that they'll actually change the temperature of their
targets in a meaningful way.
This makes them useable in combination with the thermal guns.

:cl:
balance: Completely rebalanced the temperature guns.
add: New firemode on the temperature guns
/:cl: