Catalyzing More PR-oblems (#5644)
## About The Pull Request

After some quick conversations with the admin team, and due to a lack of
player interaction with this trait, i'm proposing to reduce the cost for
the *Catalytic Inhibitor Serum* plant trait. brought by on PR #2342

May god not make me regret this.

![Alt
Text](https://tenor.com/pt-BR/view/god-help-us-all-comedian-watchmen-gif-7677780344874134701.gif)




## Why It's Good For The Game

### The Cost

20 TC cost is a very hefty value to be asked of a plant trait that
requires some good knowledge, strange seed rolls, and good prep time,
when you can just make a DIY version, with far less potency, but still
useable power, with normal cigs.

Lowering to 17 TC for an experiment should be able to tell us its value
is adequate, or if it requires further tweaking.

This also may require change in the future with more plants offering
more chems or with strange seeds being changed on how they work, which
could potentially further lower the value of this trait or increase it
depending on the changes done to the rest of the department.

![Alt
Text](https://tenor.com/pt-BR/view/guy-explaining-meme-gif-24634684.gif)
(me yapping about this)

### Fun and Suicide

Although its effect, being delaying chemical reactions from harvest to
only upon squash or thrown with the Liquid Contents trait, is indeed
strong, it's still more often than not dangerous for the botanist who
uses it.

Specially because if you add traits like Dank Vesicles and Densified
Chemicals, you get plants with so much chemicals, that you may end up
hitting yourself, like or not, with those reagents.

The only real "healthy" application in this case for its user is own
lower chemical volumes, and even so, 20 TC is asking a lot just for the
ability to throw something and its effect happen later.


![image](https://github.com/user-attachments/assets/f55e6c82-1391-4610-b50b-41300be0b89f)
( Image from our discord of what will usually happen even if you use
this Trait on your plants )

### History Vs Reality

Brought upon on PR #2342, this feature was included with a strong degree
of caution, which is very good.

Now however time has proven that being able to throw things and them
react later shouldn't cost you entire traitor budget.

17 TC might still be a lot, but with the current workings of strange
seeds, caution is still advised, even if it keeps not being used.

If this is sadly proven right, waiting for a Strange Seeds balance
change to only then reduce the cost further, might be a good path to
take in the future.


![image](http://tenor.com/pt-BR/view/hmm-michael-scott-the-office-gif-10493062.gif)
(me going into botany and hoping for things to not go crazy)

## Changelog

:cl:
balance: Reduced the cost of Catalytic Inhibitor Serum from 20 TC to 17
TC
/:cl:

---------

Co-authored-by: Lucy <lucy@absolucy.moe>