Convert fluffy tongue to a component, make Pro-tagonista Syndrome inflict fluffy tongue (#5667)
## About The Pull Request

this refactors fluffy tongue to be a component (with
`COMPONENT_DUPE_SOURCES`), rather than duplicating the same
registersignal between the quirk and the cyborg "upgrade".

i've also used this as a chance to change pro-tag syndrome inflict
fluffy tongue upon the victim, as a more punishing effect for an
otherwise often overpowered symptom, bc appending "Nyaa~" to the end of
sentences 20% of the time (often with weird double-spacing) is kinda
underwhelming.

## Why It's Good For The Game

better code and "balancing"

## Changelog
:cl:
refactor: Refactored some code relating to fluffy tongue.
balance: Pro-tagonista Syndrome now inflicts the victim with fluffy
tongue, instead of just randomly appending an poorly spaced "Nyaa~" to
20% of their messages.
/:cl: