Research designs touchup (#5662)
## About The Pull Request
- Recategorization of a decent portion of research designs, 
- Adds Multicell chargers to Advanced Engineering Designs
- Alien Tools are printable in exofabricators like Alien surgery tools
- Alien engineering tools are no longer deconstructable for alien
biotech

## Why It's Good For The Game
### Recategorization
Design paths are a bit of a mess. Equipment designs are subcategorized
as tools and machine designs are defined as computers and vice versa for
both of them. Genetics designs are in medical designs. The list goes on
and on. This tries to place them in more sensible locations while
cleaning up some ugly code like mech designs.

This will not effect anything about where is printable and how, just its
subcategorization in tabs.

Notable examples include...

- A mech LMG is no longer considered a mining tool. 
- "Boards name" (Computer Board) (Computer board) will no longer happen
- Synthetic organs are seperated from tier 1 organs
- Savannah Mechs now have listed supported equipment.
- NIFSofts, heat/cold packs, crutches, and Conveyer sorters now have
catigorzation

### Multi cell chargers
No strong reasoning for this one. I found it not in any design research
at all in the game despite have a design in the code. Its neat enough so
I added it to the advanced engineering. The wall mounted version is
still orderable via cargo on pennys on the dime so its not game
breaking.
### Alien engi tools for Exolathes
With Roboticists being the dedicated ripperdocs for anything synthetic.
They would benefit from these tools and their x10 speed in surgery. Akin
to how #2415 gave them alien surgery tools. The only one with a
mechicanical extra feature is the multitool, but its somethin I want to
see more even if it is a rare research.
### Alien engi tools no longer count to alien biotech.
Alien biotech shows up alot more than alien engineering tech, partly I
say the reason is alien surgery tools dont count for alien engineering,
but the inverse is true. This just makes it so that alien engineering
gets researched a bit more when a spare crowbar is found.
## Changelog
:cl:
add: Alien tools are printable in mechfabs
add: Multirack chargers are now researchable.
qol: Reorganized Research design categorization.
balance: Alien engineering tools are now no longer useable for alien
biotech.
fix: Mech LMG's are no longer considered a mining mech design.
/:cl: