Abandoned Containment Site Space ruins  (#4364)
## About The Pull Request

This adds a new space ruin (some sort of containment/research site)
consisting of zombies, gamer loot, and a miniboss. This is for the space
ruin competition thing.


![abandonedSite](https://github.com/user-attachments/assets/c0b765ad-8415-4ab2-bd67-5e009a1466f2)

## Why It's Good For The Game

It adds one more space ruin to the few out there in space. It is also
somewhat difficult to do so the players who want a challenge would enjoy
it.

## Changelog

:cl:
add: Added a new space ruin map
add: Added a new mob called "The Plague Doctor" which serves as a
miniboss for the ruins
add: Added more variants of the simple mob zombies
add: Added new outfits for the zombies and corpses
add: Added different variants of an ID
add: Added more corpses
add: Added "FACTION_PLAGUE" faction.
/:cl:

---------

Co-authored-by: ThePooba <81843097+ThePooba@users.noreply.github.com>