Charlie Station's equipment access ID can now open the on-station secure engineering crate (#5713)
## About The Pull Request
Fixes #5683 - it turns out the crate requires `ACCESS_AWAY_ENGINEERING`,
but the card never had it.

This also includes a tiny part of
https://github.com/tgstation/tgstation/pull/75259. Specifically, I
brought over the typepath and ID name changes to the equipment ID, as
the card is now more of a general equipment access, rather than
something specific to APCs. I opted not to bring over the change that
brings over `ACCESS_AWAY_GENERAL`, as I feel that makes this card a
replacement for the proper engineering IDs (something which it's not
meant to be).

Additionally, I replaced the two ID cards items on Charlie Station, with
more appropriate ID card items. This shouldn't result in any functional
change (aside from the changes above), as the two share the same ID
trim.

## Why It's Good For The Game
Bugfixes, and less getting in the way of a fun experience.

## Changelog

:cl: MichiRecRoom
fix: Charlie Station's equipment access ID now has access to open the
secure engineering crate that's contained in the Beta Station Engine
area.
/:cl: