Thematic obsessions (#5539)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds more associated traits for all of service to be obsessions.

- Extroverts - Bartender
- introverts - Curator
- Gourmand - Cook
- Smoker - Botanist
- Depression - Physiologist
-  Jailbird - Lawyer 
- Snob - Janitors 
- Bald - Barber  (I have no regrets on this one)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Expansion on a preexisting system. Clowns Mimes and chaplains are
already increased targets of obsessions based on their traits. Expands
it to the rest of service. Giving you thematic targets that you at least
have thematic, and mechanical reasons to be hanging out with.

As Service is the heart of the station, being jobs everyone sees and are
often in very public places. Making it less annoying to figure out how
to break into xenobiology as a medical doctor as your target stays there
all shift.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: All of service now will now have people with fanatical delusions of
'love' about them. Instead of just mime, clown, and chaplain.
code: Deleted a single instance of two TRAIT_UNCONVERTABLE being defined
in the code.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->