Remove some ai voxlines that I can't see being used in a non-rule-breaking way (#5658)
## About The Pull Request
This PR attempts to be a subset of removals taken from #4254, but only
those that would only be usable in a rule-breaking context.

The list of words that were removed are as follows:
* `abortions` - There are some legitimate non-political uses of this
word - but I can't see any reasonable AI player using this word due to
its strong connections to politics. Plus, if we're talking about
abortion in a funny spaceman game, it's very likely the round has
devolved into real-world politics. (EDIT: I was notified that
`abortions` was used to refer to xeno larvae inside people - but I'm
still iffy about it, so I'm going to defer to the admins.)
* `coomer` - Sexual word, which will almost certainly be a rule 3
violation whenever used.

Of note, I left certain words in that I wasn't sure about:
* `cunt`, `wanker` - While these can be sexual, they could also be used
by the AI to insult someone. I would hope that most players have enough
common sense to only use it as an insult.
* `woody` - This was likely added as a reference to "Woody's Got Wood" -
but the word itself can be used to describe a wood-like area.

Other words that were deleted in the original PR (which included `feet`,
for some reason) are left undeleted here, due to not being rulebreaking.

## Why It's Good For The Game
Cleans up some words that would go unused, or would only be usable in a
rule-breaking context.

## Changelog

:cl:MichiRecRoom
del: A couple words have been removed from the AI's vox repertoire -
these words were only usable in rule-breaking contexts anyways.
/:cl: