Unholy Select Equipment (#5451)
https://github.com/user-attachments/assets/0dcadb6a-59ad-441d-aad7-188ed2e41958


Example above, except its now a dropdown lol.

There's a bug with the preview not working, but I can't fix it because
my brain is air.

Awaiting a custom animation (I can't sprite, plupa is doing one <3, for
now it's just the reverse of the portal to hell smite)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds unholy select equipment for admins.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
admin: You can now choose unholy as an option in select equipment,
spawning you from hell instead.
/:cl:
- [ ] Needs sprite
- [x] Needs sound
<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->