[BOUNTY] Adds random mutations to DNA disk maints loot (#5690)
## About The Pull Request

Coded for a bounty.

This makes it so DNA data disks that spawn in maints will always have a
single random mutation.

The mutation quality (positive/neutral/negative) will be picked first,
so there won't be any bias based on how many mutations of a given
quality exist.
However, the weight for positive mutations is slightly higher than
neutral or negative mutations (5 vs 3)

There is a 70% chance that a random, applicable non-stabilizer
chromosome will be applied to the mutation. If one was not applied, then
there's a 15% chance that a stabilizer chromosome will be applied.

Almost any mutation can spawn, with a few specific exceptions:
- X-ray Vision
- Laser Eyes
- Monkified
- Unstable DNA
- The changeling variant of Chameleon

## Why It's Good For The Game

Encourages interaction with the geneticists, as you'll need a dna
console to actually check and use the mutations, and perhaps makes
things more interesting overall.

## Changelog
:cl:
add: DNA data disks that spawn in maints will now have a single random
mutation on them, sometimes with a chromosome applied.
qol: Saving a mutation from a disk to DNA console will now properly mark
that mutation as discovered if it hasn't already been discovered.
/:cl: