[PORT] Fix some runtimes in mutation code (#5796)
## About The Pull Request
This PR ports the following PRs from tgstation:
* tgstation/tgstation#79829
* tgstation/tgstation#89805

which fix some runtimes with mutations. One of these runtimes was
actually encountered on our fork - see #5768 - which is what spurred me
to fix them.

## Why It's Good For The Game
Fixes runtimes. Fixes #5768.

## Changelog

:cl:MichiRecRoom, vinylspiders
fix: (vinylspiders) fixed a race condition with mutations
fix: (MichiRecRoom) Fixes a runtime with the radioactivity mutation when
it spawns in a random data disk.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>