fix: don't dementor admins when deadmining (#5812)
## About The Pull Request

This removes two lines from the admin holder so that when the admin
deadmins, it doesn't remove their mentor status. They can always
deactivate it in the mentor tab if they don't want to mentor, but this
just makes it so that they **always** have a mentor datum.
## Why It's Good For The Game

yes

## Changelog
:cl:
fix: don't dementor admins when deadmining
/:cl: