Lowers disabler weak disabler beam debilitated mod (#5794)
## About The Pull Request
takes /obj/projectile/beam/disabler/weak debilitate_mult from 1.5 to 0.5
(used by mechs and disabler smgs)

## Why It's Good For The Game
Weak beams are supposed to be plentiful, but weak. However currently
they are stronger than a normal disabler, being a 25 energy cost cheaper
and doing 15 stamina damage.
However having a mod of 1.5 in the case of the disabler SMG (the mech
gun shoots plenty of them at once and is mostly the same case), it fires
one bullet every 0.15 seconds shooting 6 in a second, this gives it a
debilitated mod of 0.45 for every second it fires doing 5 more stamina
damage per second on EVERY shot, doing 30 damage a shot after 3 seconds,
stam critting well before this. Not to mention if you are more resistant
to disablers this hits EXTREMELY hard as it scales way way higher and no
matter how many stamina regen chems unless you are fully stun immune can
easily drop you.

Dropping this to 0.5 takes this from 5 more stamina damage per second to
roughly 2 more stamina per second, which is how it was balanced to be.
double checking this math all in a moment its slightly rounded and not
exact some of the numbers being SLIGHTLY lower or higher.


Honestly I am considering removing its debilitated multiplier because it
caps out at 20, and since it shoots so fast if you happen to be any kind
of resistant it hits you extra hard doing 180 stamina damage a second
once it maxes out. Basically full countering any kind of stamina
resistance, could also increase its cost or give it a negative mod or
something, we'll see but I think this is good enough.

## Changelog

:cl:
balance: disabler smgs now have a 0.5 debili mod instead of a 1.5
/:cl: