Mentor cloak no longer has mechanical effects (#5782)
## About The Pull Request
Removes code that harms non mentors and gives them harmful status
effects, damage, and lightning bolts them when grabbing or wearing a
mentor cloak.

## Why It's Good For The Game
This is a cloak in the loadout, it should not have built in meta checks
on if you are a mentor or not, nor punish people for wearing it as a non
mentor. While it has a check for changelings, it doesn't have one for
paradox clones, nor can it account for the many reasons you may want to
wear someones clothing to impersonate them as a antag or non antag, in
general. Or if you just want to wear their clothes.

Frankily this bit never should've made it in and I imagine it was only
merged because none of us were aware of it.

## Changelog

:cl:
del: mentor cloak no longer cares if you are a mentor
/:cl: